pigg/requirements.txt
amqp==2.5.2
anyjson==0.3.3
astroid==2.3.3
billiard==3.6.3.0
celery==4.4.1
certifi==2019.11.28
chardet==3.0.4
Django==2.0
dnspython==1.16.0
entrypoints==0.3
eventlet==0.25.1
greenlet==0.4.15
idna==2.9
importlib-metadata==1.5.0
isort==4.3.21
kazoo==2.5.0
kombu==4.6.8
lazy-object-proxy==1.4.3
logger==1.4
lxml==4.5.0
mccabe==0.6.1
monotonic==1.5
mysqlclient==1.4.6
pika==0.13.1
prometheus-client==0.7.1
pyClamd==0.4.0
pycodestyle==2.5.0
pyflakes==2.1.1
PyMySQL==0.9.3
pyserial==3.4
pytz==2019.3
PyYAML==5.3
redis==3.4.1
six==1.14.0
supervisor==4.1.0
tabulate==0.8.6
typed-ast==1.4.1
urllib3==1.25.8
vine==1.3.0
wrapt==1.11.2
zipp==3.1.0